backend-python/requirements.txt
flask
google-cloud-monitoring
google-cloud-logging
pandas
numpy
mysql-connector-python
